requirements.txt
mss
opencv-python
numpy
python-chess
stockfish
PyQt6
